pyproject.toml
[tool.poetry]
name = "photo-mosaic"
version = "0.1.0"
description = "校章モザイクアート生成アプリケーション"
authors = ["sion908 <sionn908@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
litestar = {extras = ["jinja", "standard"], version = "^2.0.0"}
pillow = "^10.0.0"
numpy = "^1.24.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false